+++ packages/react/src/twilio/components/TwilioRoomPage.tsx
@@ -1,5 +1,6 @@
 import type { BlameDictionary } from "@amfa-team/user-service";
 import React, { useCallback } from "react";
+import type { ReactElement } from "react";
 import { useJoin } from "../../api/useApi";
 import RoomPage from "../../components/RoomPage/RoomPage";
 import type { IRoom } from "../../entities/Room";
@@ -11,17 +12,23 @@ interface TwilioRoomPageProps {
   token: string;
   roomName: string;
   onRoomChanged: (roomName: string) => void;
+  onHangUp: () => void;
   blameDictionary: BlameDictionary;
+  helpButton: ReactElement;
+  featuresViewerButton: any;
 }
 
 function TwilioRoomPage(props: TwilioRoomPageProps) {
   const {
     token,
     onRoomChanged,
+    onHangUp,
     spaceId,
     userJwtToken,
     roomName,
     blameDictionary,
+    helpButton,
+    featuresViewerButton,
   } = props;
   const { data } = useConnectTwilioRoom(token);
   const { join, isJoining } = useJoin(spaceId, true, roomName);
@@ -45,8 +52,11 @@ function TwilioRoomPage(props: TwilioRoomPageProps) {
     <RoomPage
       room={room}
       onShuffle={onShuffleClicked}
+      onHangUp={onHangUp}
       isJoining={isJoining}
       blameDictionary={blameDictionary}
+      helpButton={helpButton}
+      featuresViewerButton={featuresViewerButton}
     />
   );
 }